requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
Django==5.2.5
django-crispy-forms==2.4
Pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2

# AI News Generation Dependencies
openai==1.57.0
python-decouple==3.8
requests==2.32.3
beautifulsoup4==4.12.3

# Production Dependencies
gunicorn==22.0.0
psycopg2-binary==2.9.9
boto3==1.35.39
django-storages==1.14.4
whitenoise==6.8.2
